Refactor and rename cli-e2e tests
