+++ package.json
@@ -31,7 +31,7 @@
     "eslint-plugin-react-hooks": "^4.2.0",
     "eslint-plugin-tsdoc": "^0.2.7",
     "husky": "^4.2.5",
-    "jest": "^26.4.2",
+    "jest": "^26.6.3",
     "lerna": "^3.22.0",
     "lint-staged": "^10.2.11",
     "prettier": "^2.0.1",
