+++ packages/frontity/package.json
@@ -24,11 +24,13 @@
     "start": "node dist/src/cli/index.js",
     "dev": "ts-node src/cli/index.ts",
     "ts": "ts-node",
-    "test": "../../node_modules/.bin/jest --watch",
-    "test:ci": "../../node_modules/.bin/jest --colors --ci --coverage",
-    "test:inspect": "node --inspect ../../node_modules/jest/bin/jest.js --watch --no-cache --runInBand",
-    "test:e2e-cli-full": "../../node_modules/.bin/jest --testRegex /packages/frontity/e2e/cli.full.ts",
-    "build": "tsc --project ./tsconfig.build.json && cp -a templates dist/templates",
+    "test": "jest --watch",
+    "test:ci": "jest --colors --ci --coverage",
+    "test:inspect": "node --inspect node_modules/bin/jest --watch --no-cache --runInBand",
+    "test:cli-e2e:filesystem": "jest cli-e2e/filesystem",
+    "test:cli-e2e:docker": "jest cli-e2e/docker",
+    "build": "tsc --project ./tsconfig.build.json && npm run copyfiles",
+    "copyfiles": "copyfiles templates/* dist",
     "build:watch": "tsc --project ./tsconfig.build.json --watch",
     "prepublish": "npm run build"
   },
@@ -76,7 +78,9 @@
     "@types/shelljs": "^0.8.8",
     "@types/stack-trace": "^0.0.29",
     "@types/tar": "^4.0.4",
+    "copyfiles": "^2.4.1",
     "execa": "^5.0.0",
+    "jest": "^26.6.3",
     "lodash.omit": "^4.5.0",
     "react-test-renderer": "^17.0.1",
     "ts-node": "^9.1.1"
